create SemModuleScreen
