--- components/SemesterComponent.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { StyleSheet, Text, View, TouchableOpacity, SafeAreaView } from 'react-native';
-
-const SemesterComponent = () => {
-  return (
-<View>
-    <Text>HI</Text>
-</View>
-  )
-}
-
-export default SemesterComponent
